+++ src/test/java/blackjack/domain/BlackjackTest.java
@@ -1,10 +1,8 @@
-package blackjack;
+package blackjack.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import blackjack.domain.Blackjack;
-import blackjack.domain.card.CardValue;
-import blackjack.domain.card.Cards;
 import blackjack.domain.common.Name;
 import blackjack.domain.common.Names;
 import blackjack.domain.player.Dealer;
