+++ src/test/java/blackjack/domain/common/NameTest.java
@@ -1,4 +1,4 @@
-package blackjack;
+package blackjack.domain.common;
 
 import static org.assertj.core.api.Assertions.assertThatCode;
