+++ src/test/java/blackjack/domain/player/PlayerTest.java
@@ -1,9 +1,5 @@
-package blackjack;
+package blackjack.domain.player;
 
-import blackjack.domain.card.Card;
-import blackjack.domain.card.CardSymbol;
-import blackjack.domain.card.CardValue;
-import blackjack.domain.card.Cards;
 import blackjack.domain.common.Name;
 import blackjack.domain.player.GamePlayer;
 import blackjack.fixture.PlayerFixture;
