+++ src/test/java/blackjack/domain/result/DealerResultTest.java
@@ -1,4 +1,4 @@
-package blackjack;
+package blackjack.domain.result;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
