+++ src/test/java/blackjack/domain/result/GamePlayerResultTest.java
@@ -1,4 +1,4 @@
-package blackjack;
+package blackjack.domain.result;
 
 import static org.assertj.core.api.Assertions.assertThatCode;
